+++ engine/engine_test.go
@@ -82,40 +82,4 @@ var _ = Describe("type Engine", func() {
 			}).To(Panic())
 		})
 	})
-
-	Describe("func CommandExecutor()", func() {
-		It("returns a dogma.CommandExecutor that dispatches to the engine", func() {
-			called := false
-			aggregate.HandleCommandFunc = func(
-				_ dogma.AggregateCommandScope,
-				m dogma.Message,
-			) {
-				called = true
-				Expect(m).To(Equal(MessageA1))
-			}
-
-			err := engine.CommandExecutor().ExecuteCommand(context.Background(), MessageA1)
-			Expect(err).ShouldNot(HaveOccurred())
-			Expect(called).To(BeTrue())
-		})
-	})
-
-	Describe("func EventRecorder()", func() {
-		It("returns a dogma.EventRecorder that dispatches to the engine", func() {
-			called := false
-			process.HandleEventFunc = func(
-				_ context.Context,
-				_ dogma.ProcessEventScope,
-				m dogma.Message,
-			) error {
-				called = true
-				Expect(m).To(Equal(MessageE1))
-				return nil
-			}
-
-			err := engine.EventRecorder().RecordEvent(context.Background(), MessageE1)
-			Expect(err).ShouldNot(HaveOccurred())
-			Expect(called).To(BeTrue())
-		})
-	})
 })